adding sets
