--- activiti-spring-conformance/activiti-spring-conformance-basic-set/pom.xml
@@ -1,70 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <groupId>org.activiti</groupId>
-    <artifactId>activiti-spring-conformance</artifactId>
-    <version>7.0.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>activiti-spring-conformance-basic-set</artifactId>
-  <packaging>jar</packaging>
-  <name>Activiti Spring :: Conformance Basic Set</name>
-  <description>Activiti Spring :: Conformance Basic Set</description>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.activiti.api</groupId>
-      <artifactId>activiti-api-process-runtime</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.activiti.api</groupId>
-      <artifactId>activiti-api-task-runtime</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.activiti.api</groupId>
-      <artifactId>activiti-api-runtime-shared</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.activiti.api</groupId>
-      <artifactId>activiti-api-task-model</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.activiti.api</groupId>
-      <artifactId>activiti-api-process-model</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.activiti.api</groupId>
-      <artifactId>activiti-api-model-shared</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.activiti</groupId>
-      <artifactId>activiti-engine</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.activiti</groupId>
-      <artifactId>activiti-spring-boot-starter</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-web</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.slf4j</groupId>
-      <artifactId>slf4j-api</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.h2database</groupId>
-      <artifactId>h2</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.springframework.boot</groupId>
-      <artifactId>spring-boot-starter-test</artifactId>
-      <scope>test</scope>
-    </dependency>
-  </dependencies>
-
-</project>
