+++ activiti-spring-conformance/activiti-spring-conformance-set0/src/test/resources/processes/Process with Generic  BPMN Task.bpmn20.xml
