+++ activiti-spring-conformance/activiti-spring-conformance-set1/src/test/java/org/activiti/spring/conformance/set1/Application.java
@@ -1,0 +1,13 @@
+package org.activiti.spring.conformance.set1;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class);
+    }
+
+}
