+++ activiti-spring-conformance/activiti-spring-conformance-set1/src/test/java/org/activiti/spring/conformance/set1/security/util/SecurityUtil.java
@@ -1,0 +1,77 @@
+package org.activiti.spring.conformance.set1.security.util;
+
+import org.activiti.api.runtime.shared.security.SecurityManager;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.context.SecurityContextImpl;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.stereotype.Component;
+
+import java.util.Collection;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+@Component
+public class SecurityUtil {
+
+    private Logger logger = LoggerFactory.getLogger(SecurityUtil.class);
+
+    @Autowired
+    private UserDetailsService userDetailsService;
+
+    @Autowired
+    private SecurityManager securityManager;
+
+    public void logInAs(String username) {
+
+        UserDetails user = userDetailsService.loadUserByUsername(username);
+        if (user == null) {
+            throw new IllegalStateException("User " + username + " doesn't exist, please provide a valid user");
+        }
+        logger.info("> Logged in as: " + username);
+        SecurityContextHolder.setContext(new SecurityContextImpl(new Authentication() {
+            @Override
+            public Collection<? extends GrantedAuthority> getAuthorities() {
+                return user.getAuthorities();
+            }
+
+            @Override
+            public Object getCredentials() {
+                return user.getPassword();
+            }
+
+            @Override
+            public Object getDetails() {
+                return user;
+            }
+
+            @Override
+            public Object getPrincipal() {
+                return user;
+            }
+
+            @Override
+            public boolean isAuthenticated() {
+                return true;
+            }
+
+            @Override
+            public void setAuthenticated(boolean isAuthenticated) throws IllegalArgumentException {
+
+            }
+
+            @Override
+            public String getName() {
+                return user.getUsername();
+            }
+        }));
+        org.activiti.engine.impl.identity.Authentication.setAuthenticatedUserId(username);
+
+        assertThat(securityManager.getAuthenticatedUserId()).isEqualTo(username);
+    }
+}
