+++ activiti-spring-conformance/activiti-spring-conformance-set2/src/test/resources/processes/UserTask with Assignee.bpmn20.xml
