+++ activiti-spring-conformance/pom.xml
@@ -15,6 +15,8 @@
     <description>Activiti Spring :: Conformance Parent</description>
 
     <modules>
-        <module>activiti-spring-conformance-basic-set</module>
+        <module>activiti-spring-conformance-set0</module>
+        <module>activiti-spring-conformance-set1</module>
+        <module>activiti-spring-conformance-set2</module>
     </modules>
 </project>
